Single source of sample mapping spec is used in all places
